refactor(onboarding): move form components to a separate folder and apply FormWrapper
- Moved PersonalInfoForm, WhatFeatures, HowDidYouHear, and LinkedInUrl to a new 'forms' folder for better organization.
- Created FormWrapper component to wrap form content and applied it to the form components.
