+++ src/app/(auth)/sign-up/(onboarding)/components/forms/FormWrapper.tsx
@@ -1,0 +1,7 @@
+interface FormWrapperProps {
+  children: React.ReactNode;
+}
+
+export default function FormWrapper({ children }: FormWrapperProps) {
+  return <div className="py-10">{children}</div>;
+}
