+++ src/app/(auth)/sign-up/(onboarding)/components/forms/PersonalInfoForm.tsx
@@ -1,10 +1,11 @@
 import TextInput from "@/components/inputs/TextInput";
+import FormWrapper from "@/app/(auth)/sign-up/(onboarding)/components/forms/FormWrapper";
 
 export default function PersonalInfoForm() {
   return (
-    <div className="py-10">
+    <FormWrapper>
       <TextInput id="firstName" label="First Name" placeholder="e.g. John" />
       <TextInput id="lastName" label="Last Name" placeholder="e.g. Smith" />
-    </div>
+    </FormWrapper>
   );
 }
