+++ src/app/(auth)/sign-up/(onboarding)/components/onboardingSteps.tsx
@@ -1,9 +1,9 @@
 import WelcomeChingu from "@/app/(auth)/sign-up/(onboarding)/components/WelcomeChingu";
-import PersonalInfoForm from "@/app/(auth)/sign-up/(onboarding)/components/PersonalInfoForm";
+import PersonalInfoForm from "@/app/(auth)/sign-up/(onboarding)/components/forms/PersonalInfoForm";
 import OnboardingImage from "@/app/(auth)/sign-up/(onboarding)/components/OnboardingImage";
-import WhatFeatures from "@/app/(auth)/sign-up/(onboarding)/components/WhatFeatures";
-import HowDidYouHear from "@/app/(auth)/sign-up/(onboarding)/components/HowDidYouHear";
-import LinkedInUrl from "@/app/(auth)/sign-up/(onboarding)/components/LinkedInUrl";
+import WhatFeatures from "@/app/(auth)/sign-up/(onboarding)/components/forms/WhatFeatures";
+import HowDidYouHear from "@/app/(auth)/sign-up/(onboarding)/components/forms/HowDidYouHear";
+import LinkedInUrl from "@/app/(auth)/sign-up/(onboarding)/components/forms/LinkedInUrl";
 
 type OnboardingStep = {
   handleNext?: () => void; // Optional for steps without handleNext
